BUGFIX:  /play endpoint was slow
Before this fix, the /play API endpoint would block until the
entire duration of the requested program before returning a response
to the caller.

Now I launch a go routine to pause, play the url, and unpause.
This allows the API endpoint to return independent of the activity
of playing the URL for the duration.